Migrate to single binary layout
The previous structure of having one binary per year was a holdover
from grafting the separate repositories together. Ultimately, storing
and processing the `Solution`s the same way is preferable because it
eliminates the overhead of replicating `main()` functions everywhere
and promotes better code reuse.

Accordingly, this is a humongous list of files to shuffle around since
the whole project structure changed.